proxy_api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
httpx==0.25.1
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pyjwt==2.8.0
tenacity==8.2.3
prometheus-client==0.17.1
python-multipart==0.0.6
aiofiles==23.2.1
databases[postgresql]==0.8.0
sqlalchemy==2.0.23
starlette==0.27.0
loguru==0.7.2
